refactor: factored out BoardWindow
